Added few improvements to make MagicMirror better
Used momentjs' translations for every DateTime strings. No more need to add translations.

Added config.js to use as a central file for configurations.

Try to use Gmail's calendars and I've seen that there are some problems (with old events,
when an event has "email remembers", and with TimeZones). I'll try to solve it soon.