Replace Helvetica Neue with Open-Source Font
Helvetica Neue is a proprietary font. Redistributing it in this
way is not considered fair use. This commit removes the Helvetica
Neue fonts from the repository and instead includes Roboto. It is
written in such a way that if you copy the Helvetica Neue fonts
back into the fonts/ directory (under a `helvetica-neue` subfolder)
they will be used as the primary font instead of Roboto.