Added relative/absolute as well as 'urgency' time display on events
User is now able to pick between 'relative' or 'absolute' in their
config. When using 'absolute', the user can also set an 'urgency' time
frame (in days) which will change the display of the event time from
'absolute' to 'relative'. This allows the user to have events that
display both relative and absolute times.