Fix handling of recurring events and add monthly view
The primary fix is to the handling of recurring events.  For recurrence
rules in iCal (RRULE), it's possible to have overrides for specific
recurrences, and "exception dates" (EXDATE) for dates that a recurrence
shouldn't happen.  ical.js was ignoring EXDATE and just overwriting
RRULE entries with the specific recurrence overrides.  Here are the
fixes:
- ical.js - added a recurrences[] array to store data about the specific
override.  added an exdate[] array to story data about exception dates.
(also added dtstamp, created, and lastmodified fields just for a bit
more completeness)
- calendarfetcher.js - changed the logic to handle the new recurrence
and exdate arrays correctly.  also added the ability to force a specific
start date to make testing easier. :)
- debug.js / node_helper.js - added pass-through code to pass the start
date override to calendarfetcher

Also added an optional monthly view to the calendar module (calendar.js
/ calendar.css).