Add support for wind direction as an arrow
The abbreviations for wind directions do not work similarly in Finnish as in English.

| English         | Finnish compass direction |
|-----------------|---------------------------|
| North           | pohjoinen                 |
| Northnortheast  | _no word_                 |
| Northeast       | koillinen                 |
| Eastnortheast   | _no word_                 |
| East            | itä                       |
| Eastsoutheast   | _no word_                 |
| Southeast       | kaakko                    |
| Southsoutheast  | _no word_                 |
| South           | etelä                     |
| Southsouthwest  | _no word_                 |
| Southwest       | lounas                    |
| Westsouthwest   | _no word_                 |
| West            | länsi                     |
| Westnorthwest   | _no word_                 |
| Northwest       | luode                     |
| Northnorthwest  | _no word_                 |

At least I do not know words for the minor directions in Finnish. So the translation is a bit gibberish. In my mirror project I have solved this by adding a feature of showing the wind direction with a little arrow pointing where the wind is blowing to. So, if you'd see `3N` normally, I'd see 3 and a small arrow pointing down, as the wind is coming from north and blowing towards the observer, towards south.

With this proposed change, you could add `showWindDirectionAsArrow: true` to the current weather configuration to toggle the arrow on.