Prioritize the address provided instead of the server version
The address variable on the remote server can be set differently (e.g to localhost) than the one provided when calling node clientonly. This is especially true, if MM is running in a Docker. 
The address variable was overwritten with the one fetched from the remote server before spawning the electron application which could not find the server in return.