Added showExtraWeatherInfo option
Added this new option because I wanted to display temp and feelsLike
and nothing else. The onlyTemp option was the only way to turn off
the extra weather info above the temp, but it also blocked the feelLike.
This is a more consistent way to "build" the weather display rather than
using onlyTemp. onlyTemp still works, but it should be considered deprecated.