Add files via upload
forecast.njk - Updated precipitation logic such that if weatherObject has precipitationUnits specified, they will be displayed instead of the default logic that specifies 'in' , 'mm', etc.  This allows the new envCanada provider to use the precipitation units included in the Environment Canada forecast data. It also better reflects that Canadian forecasts tend to measure snow in 'cm' and rain in 'mm'.

weatherobject.js - A new object element is included = precipitationUnits. It is initialized as null, and can be optionally used to specify a custom unit measurement to be displayed in a forecast. The new envCanada provider uses this (see forecast.njk). If this object element is not populated with a non-null value, then default processing is used to select the correct unit to display (e.g. 'in', 'mm', etc.).

weather.js - The getScripts input has been updated to include 'providers/xml2json.min.js'. This utility function is used by the new envCanada provider to convert XML documents provided by Environment Canada into JSON structures, which are easier to work with in the provider code.