Implemented Yr as a weather provider.
The Yr documentation was quite strict on using a cache together with the expires- and if-modified since-headers. As a result the weather provider has to send and recieve HTTP-headers. We use the default User-Agent for now, even though it is on a slightly different format than the one documented by Yr.