[core] fix: allow browser globals in config files
The config checker previously only allowed Node.js globals, but since
the config file runs also in the browser context, users should
be able to access browser APIs like `document` or `window` when needed.

This was incorrectly flagged as an error by the `no-undef` ESLint rule.
The fix adds browser globals to the allowed globals in the linter config.

Fixes #3990